refactor(server): 내정보 관련 로직들을 user 에서 my 로 이동 #213
