+++ apps/server/src/services/user.service.ts
@@ -47,112 +47,3 @@ export const getOneForCheck = async (id: string) => {
 
   return user;
 };
-
-////////////////////////////////////////////////////////////////
-
-/**
- * 고객 상세정보 조회
- */
-export const getCustomerInfo = async (no: number) => {
-  const result = await prisma.$transaction(async (tx) => {
-    // 고객정보
-    const customer = await tx.customer.findUnique({
-      where: { no },
-      select: {
-        no: true,
-        id: true,
-        name: true,
-        mobileNumber: true,
-        defaultAddressNo: true,
-        createdAt: true,
-
-        // relationship
-        address: true,
-        customerCoupon: true,
-      },
-    });
-
-    // 할인정보
-    const coupon = await tx.customerCoupon.findMany({
-      where: { customerNo: no },
-      include: {
-        coupon: true,
-      },
-    });
-
-    return { customer, coupon };
-  });
-
-  return result;
-};
-
-/**
- * 주문 누적금액 조회
- */
-export const getOrderAccumulatedAmount = async (no: number) => {
-  const result = await prisma.$transaction(async (tx) => {
-    let totalPrice = 0;
-
-    const getTotalPrice = await tx.order.aggregate({
-      where: { customerNo: no },
-      _sum: { totalPrice: true },
-    });
-
-    if (getTotalPrice._sum) {
-      totalPrice = getTotalPrice._sum.totalPrice || 0;
-    }
-
-    return totalPrice;
-  });
-
-  return result;
-};
-
-/**
- * 고객정보 수정
- */
-export const update = async (data: {
-  id: string;
-  name: string;
-  mobileNumber: string;
-  createdAt: string;
-}) => {
-  const result = await prisma.$transaction(async (tx) => {
-    const updateCustomer = await tx.customer.update({
-      where: { id: data.id },
-      data: { mobileNumber: data.mobileNumber },
-    });
-
-    return updateCustomer;
-  });
-
-  return result;
-};
-
-/**
- * 고객정보 상세 조회(민감정보 조회 전용)
- * email, pw, refreshToken 등
- */
-export const findCustomerDetail = async (id: string) => {
-  const result = await prisma.$transaction(async (tx) => {
-    const customer = await tx.customer.findUnique({ where: { id } });
-
-    return customer;
-  });
-
-  return result;
-};
-
-/**
- * 고객정보 수정 : 비밀번호 변경
- */
-export const updatePw = async (no: number, hashedPw: string): Promise<void> => {
-  await prisma.$transaction(async (tx) => {
-    await tx.customer.update({
-      where: { no },
-      data: {
-        pw: hashedPw,
-      },
-    });
-  });
-};
